write_bmp.hpp -> bmp.hpp: 新增读取 bmp 文件的功能 示例 BMPFile f("file.bmp"); f.info(); f.output("file.bak.bmp");
